Moved the error component to the shared folder. ErrorComponent is now exported with his module. Updated a few files import - WIP
