+++ angular-client/src/app/app.module.ts
@@ -4,23 +4,25 @@ import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import { HttpModule } from '@angular/http';
 
 import { AppComponent } from './app.component';
-import {AuthenticationComponent} from "./modules/authentification/authentication.component";
-import {HeaderComponent} from "./shared/components/header/header.component";
-import {Routing} from "./app.routing";
-import {TeamsComponent} from "./modules/teams/teams.component";
-import {ChallengesComponent} from "./modules/challenges/challenges.component";
-import {HomeComponent} from "./modules/home/home.component";
-import {JudgeComponent} from "./modules/judge/judge.component";
-import {SignupComponent} from "./modules/authentification/signup.component";
-import {SigninComponent} from "./modules/authentification/signin.component";
-import {ChallengeComponent} from "./modules/challenges/challenge.component";
-import {ChallengeInputComponent} from "./modules/challenges/challenge-input.component";
-import {ChallengeListComponent} from "./modules/challenges/challenge-list.component";
-import {AuthService} from "./modules/authentification/auth.service";
-import {ErrorService} from "./modules/errors/error.service";
-import {ErrorComponent} from "./modules/errors/error.component";
-import {AccountComponent} from "./modules/authentification/account.component";
-import {CanActivateViaAuthGuard} from "./modules/authentification/can-activate";
+import {AuthenticationComponent} from './modules/authentification/authentication.component';
+import {HeaderComponent} from './shared/modules/header/header.component';
+import {Routing} from './app.routing';
+import {TeamsComponent} from './modules/teams/teams.component';
+import {ChallengesComponent} from './modules/challenges/challenges.component';
+import {HomeComponent} from './modules/home/home.component';
+import {JudgeComponent} from './modules/judge/judge.component';
+import {SignupComponent} from './modules/authentification/signup.component';
+import {SigninComponent} from './modules/authentification/signin.component';
+import {ChallengeComponent} from './modules/challenges/challenge.component';
+import {ChallengeInputComponent} from './modules/challenges/challenge-input.component';
+import {ChallengeListComponent} from './modules/challenges/challenge-list.component';
+import {AuthService} from './modules/authentification/auth.service';
+import {ErrorService} from './shared/services/error.service';
+import {AccountComponent} from './modules/authentification/account.component';
+import {CanActivateViaAuthGuard} from './modules/authentification/can-activate';
+
+import { ErrorModule } from '@shared-modules';
+
 
 @NgModule({
   declarations: [
@@ -36,7 +38,6 @@ import {CanActivateViaAuthGuard} from "./modules/authentification/can-activate";
     JudgeComponent,
     SignupComponent,
     SigninComponent,
-    ErrorComponent,
     AccountComponent
   ],
   imports: [
@@ -44,9 +45,15 @@ import {CanActivateViaAuthGuard} from "./modules/authentification/can-activate";
     FormsModule,
     HttpModule,
     Routing,
-    ReactiveFormsModule
+    ReactiveFormsModule,
+
+    ErrorModule
   ],
-  providers: [AuthService, ErrorService, CanActivateViaAuthGuard], //provide to all of the application
+  providers: [
+    AuthService,
+    ErrorService,
+    CanActivateViaAuthGuard
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
