+++ angular-client/src/app/modules/authentification/account.component.ts
@@ -1,5 +1,5 @@
-import { Component } from "@angular/core";
-import { AuthService } from "./auth.service";
+import { Component } from '@angular/core';
+import { AuthService } from './auth.service';
 
 @Component({
   selector: 'app-account',
