+++ angular-client/src/app/modules/authentification/auth.routes.ts
@@ -1,9 +1,9 @@
-import {Routes} from "@angular/router";
-import {SignupComponent} from "./signup.component";
-import {SigninComponent} from "./signin.component";
+import {Routes} from '@angular/router';
+import {SignupComponent} from './signup.component';
+import {SigninComponent} from './signin.component';
 
 export const AUTH_ROUTES: Routes = [
-  { path: '', redirectTo: 'signup', pathMatch:'full' },
+  { path: '', redirectTo: 'signup', pathMatch: 'full' },
   { path: 'signup', component: SignupComponent },
   { path: 'signin', component: SigninComponent },
 ];
