+++ angular-client/src/app/modules/authentification/auth.service.ts
@@ -1,9 +1,9 @@
-import {User} from "./user.model";
-import {Injectable} from "@angular/core";
-import {Http, Headers, Response} from "@angular/http";
+import {User} from '../../shared/models/user.model';
+import {Injectable} from '@angular/core';
+import {Http, Headers, Response} from '@angular/http';
 import 'rxjs/Rx';
-import {Observable} from "rxjs/Observable";
-import {ErrorService} from "../errors/error.service";
+import {Observable} from 'rxjs/Observable';
+import {ErrorService} from '../../shared/services/error.service';
 
 @Injectable()
 
@@ -13,9 +13,9 @@ export class AuthService {
 
   constructor(private http: Http, private errorService: ErrorService){}
 
-  signup(user: User){
+  signup(user: User) {
     const body = JSON.stringify(user);
-    const headers = new Headers({'Content-Type':'application/json'});
+    const headers = new Headers({'Content-Type': 'application/json'});
     return this.http.post(`${this.API}/user`, body, {headers: headers})
       .map((response: Response) => response.json())
       .catch((error: Response) => {
@@ -35,15 +35,15 @@ export class AuthService {
       });
   }
 
-  logout(){
-    localStorage.clear()
+  logout() {
+    localStorage.clear();
   }
 
-  isLoggedIn(){
+  isLoggedIn() {
     return localStorage.getItem('token') !== null;
   }
 
-  getToken(){
+  getToken() {
     return localStorage.getItem('token');
   }
 }
