+++ angular-client/src/app/modules/authentification/signin.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from "@angular/core"
 import {FormControl, FormGroup, Validators} from "@angular/forms";
-import {User} from "./user.model";
+import {User} from "../../shared/models/user.model";
 import {AuthService} from "./auth.service";
 import {Router} from "@angular/router";
