+++ angular-client/src/app/modules/authentification/signup.component.ts
@@ -1,8 +1,8 @@
-import {Component, OnInit} from '@angular/core';
-import {FormControl, FormGroup, Validators} from '@angular/forms';
+import { Component, OnInit } from '@angular/core';
+import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
-import {AuthService} from './auth.service';
-import {User} from './user.model';
+import { AuthService } from './auth.service';
+import { User } from '@models';
 // import { PasswordValidation } from "./password-validator";
 
 @Component({
