+++ angular-client/src/app/modules/home/home.component.ts
@@ -1,4 +1,4 @@
-import { Component } from "@angular/core";
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-home',
