+++ angular-client/src/app/modules/judge/judge.component.ts
@@ -1,4 +1,4 @@
-import { Component } from "@angular/core";
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-judge',
