+++ angular-client/src/app/shared/models/error.model.ts
