+++ angular-client/src/app/shared/models/user.model.ts
