+++ angular-client/src/app/shared/modules/errors/error.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { HttpModule } from '@angular/http';
+
+import { ErrorComponent } from './error.component';
+
+
+@NgModule({
+  imports: [
+    CommonModule,
+    HttpModule
+  ],
+  declarations: [ ErrorComponent ],
+  exports: [ ErrorComponent ]
+})
+
+export class ErrorModule { }
