+++ angular-client/src/app/shared/modules/header/header.component.ts
